scipy integerate
### **What is `scipy.integrate`?**

In simple words, integration is like calculating the area under a curve, right? Matlab: If you have a function, and you want to find out the total area from one point to another on its curve, you integrate that function.

Now, in **SciPy**, we have a module called **`integrate`**, which is *very* useful for solving both simple and complex integration problems. Whether it's single-variable integrals, multi-variable integrals, or solving differential equations—this is the toolkit you need!

Co-Authored-By: 왕뚜껑맛있땅 <7420085+JIwoo@users.noreply.github.com>
Co-Authored-By: Hyoeun  <137873124+hyoeunjoo@users.noreply.github.com>
Co-Authored-By: jiwooverse <184836366+jiwooverse@users.noreply.github.com>
Co-Authored-By: Jayson <89952042+jsoonworld@users.noreply.github.com>
Co-Authored-By: VIMAL KUMAR <vimalkumarji112@gmail.com>